refactor: move `get_excel_sheet_names()` into `functions.py` for better unit testing
